Add --aggr flag
Using this flag one can sum all states and use the percentage flags to
check whether the percentage of (sum of all states / total_results) is over some
threshold. This works well on large installations where you might have a
lot of (flapping) alerts but you just care if the majority of your
services are in an OK state.

If the flag is used, status line includes the total percentage of
problems as well.

UNKNOWNs are treated a bit special, and exitcode works the same way as
without the --aggr flag.